Osetrovanie na strane servera.

+++ App/Controllers/HomeController.php
@@ -8,6 +8,7 @@ use App\Models\Orderproducts;
 use App\Models\Orders;
 use App\Models\Products;
 use App\Models\User;
+use mysql_xdevapi\Exception;
 
 class HomeController extends AControllerBase
 {
@@ -19,7 +20,11 @@ class HomeController extends AControllerBase
 
     public function delete(): Response
     {
-        $this->app->getAuth()->deleteUser();
+        if (isset($_SESSION['userid']) && $_SESSION['userid'] > 0) {
+            $this->app->getAuth()->deleteUser();
+        } else {
+            throw new \Exception("Nemáte oprávnenie na túto akciu.");
+        }
         return $this->redirect($this->url("home.index"));
     }
 
@@ -40,15 +45,27 @@ class HomeController extends AControllerBase
 
     public function orders(): Response
     {
+        if (!isset($_SESSION['userid']) || $_SESSION['userid'] <= 0) {
+            throw new \Exception("Pre zobrazenie objednávok sa musíte prihlásiť.");
+        }
         $user = $_SESSION['userid'];
         $data = Orders::getAll('user_id = ?', [$user]);
+        if (!$data) {
+            throw new \Exception("Žiadne objednávky neboli nájdené.");
+        }
         return $this->html($data);
     }
 
     public function ordersinfo(): Response
     {
         $orderid = $_GET['id'];
+        if (!isset($_SESSION['userid']) || $_SESSION['userid'] <= 0) {
+            throw new \Exception("Pre zobrazenie informácií o objednávke sa musíte prihlásiť.");
+        }
         $data = Orderproducts::getAll('order_id = ?', [$orderid]);
+        if (!$data) {
+            throw new \Exception("Informácie o objednávke neboli nájdené..");
+        }
         return $this->html($data);
     }
 }
